requirements.txt
fastapi
uvicorn
tensorflow
tensorflow-hub
numpy
opencv-python-headless
pillow
requests
